Inverse transform
Added the inverse transform which is built on the same recursive radix-2 FFT function as the forward transform. This required changing the input to a vector of complex numbers which then requires a copy of the input for the forward transform making it a little less efficient.

Testing shows that vec = IFFT(FFT(vec)) as expected for size 8 inputs. However,  this fails for size 16 inputs which reveals the issue only arises for FFTs larger than 8.