FFT fix and test cases
The previous issue with the FFT implementation arose because the binary tree was traversing the input and output at the same time which would result in a butterfly being calculated using values that had not been set yet from the initial 2-point FFTs on the input. This was solved by traversing the input and output binary trees separately.

For the input binary tree, the traversal was replaced with a simple loop over the pairs of leaves as only the leaves matter, it is a waste of time to traverse the inner nodes in the array. For the output binary tree, its traversal was also switched to use for loops to remove recursion entirely.

All tests are now passed. More test cases were added for larger FFTs and for inputs that are no powers of 2. This showed that there is some ambiguity in how the functions should work if the given input is a larger size than the given n. It seems the transform should ignore all values after the first n, but it currently does not. More work is needed to clarify function behavior.