pybind11 configuration
Configuration for CMake to detect the external pybind module by directing it to the CMake config file included with pybind11

A find module is used as an intermediate before leading CMake to the config as way to separate the configuration of the local directory where pybind11 is found. Perhaps there is a better way to do this, but vscode does not pick up environment variables set in my python venv--I think vscode runs CMake outside the venv.