initial pybind setup
Using pybind's built in translation between C++ types and Python types results in the bound transforms not passing their output to Python because input values are passed by copy.

To pass by reference, a consistent type must be used and it makes much more sense to have the C++ interface with python types than to have Python interface with C++ types given that relatively seamless interfacing with numpy is desired.

numpy arrays cannot be resized which requires either a translation layer that then call the transforms which are built on resizable vectors or refactoring the transforms to use std::arrays instead of vectors.