Removed transform size from template parameters
Transform size is now provided via constructor

Following the pattern of forbidding partially initialized objects, transform size is now provided via constructor. Further, it must be computed before reaching the body of the constructor so the code for computing the FFT size was moved to a static helper function which can be called within the initializer list of constructors. This call needs to be made by both the FFT and DCT (cannot set size of DCT in parent constructor based on the contained FFT size, need to call the parent constructor first) so it has been moved to the Abstract_Transformer template. This does mean the code will be duplicated for every parameterization of the template which perhaps justifies moving it to a static helper class.

This commit also contains an experiment with using unsigned types to match STL containers instead of manually converting container size. This will likely be short lived giving a switch to using spans for transform input and output is likely the next phase. The goal is for the view of spans to provide a view that can take a range of containers such as STL containers or python types without requiring a copy.