Method signatures take spans now
Changed the use of vectors to spans for calls to transform and inverse to increase flexibility on what sorts of objects can be passed. This was motivated by the need to make a copy when converting a Python array to a vector.

The switch to unsigned indices has been undone. int is much more convenient and immediate to use so it is not worth risking unexpected behavior from unsigned values unless there is a very good reason to use unsigned values.

This change does mean that the transform and inverse methods can no longer change the size of a vector being written to so any calls where too small a span is provided are rejected. Following single responsibility, this is probably an improvement.

Added an option to have a vector returned when no output span is provided to transform or inverse. It can be much more convenient to avoid having to size the output array before a function call. This was implemented as an overload in the base class. Due to name hiding, a using statement needs to be added in each inheriting class. It is worth reconsidering the decision to use overloading in the future. Copy elision should eliminate any performance  hit from the copy.

All tests are passed on the C++ end, but pybind11 does not have built in support for spans so the python end still does not work. The API changes motivated by pybind11 are starting to settle so those are almost ready to be pulled into the main branch.