Python binding fix
The python bindings now work and can be used to properly perform the transforms within python. For now, only the DCT_2 has been bound.

Using a C++ span for reading numpy data without needing to copy it worked. There is currently no support for writing the output directly to a provided buffer and currently no type checks are performed on top of those that pybind11 performs.

Current idea for handling python bindings is to create an inheriting class for each transform class that adds functions which takes numpy arrays as input and delegate to the parent class (the transform) to avoid having any pybind code mixed in with code that implements transforms.