Initial setup for image processing
Created lightweight classes for passing around views of image data built on spans.
The intent is for the views to be trivially copyable and to provide convenience methods for accessing image data that is in contiguous memory.
This is to ultimately allow the views to apply to numpy arrays to allow processing of data from python without a copy.

The 2D naming is likely to change given that a color 2D image has multiple color channels so it is mathematically 3 dimensional.

The current use of the abstract_transformer is also somewhat unsatisfying as it requires data to be passed as span. This can make sense, but it seems preferable to have inputs and outputs wrapped as "Multichannel_Matrix" (current naming) views to pass input and output shape along with the data.
Additional methods could be added to support calling transform and inverse on non-span objects or the storage class could be passed to the template as another parameter.

Channel is implemented as a template parameter unlike shape given a single transform can make sense for many image shapes, but only certain numbers of image channels.