YCbCr conversion setup
Created the setup for an initial YCbCr transformer class and its testing. This motivated significant changes in the 2d (tentative naming) code with the use of const members abandoned for the matrix structs/classes.
The goal of the matrix structs is to provide views. Const members of a struct/class prevent a variable of that type from being assigned to. This is unnaceptable for a view object which is meant to be easily passed around.
To adjust, internal variables are now private and follow the "name_" convention and accessed via "name() const" methods.

The base Abstract_Transformer class was modified to accept a template template to allow different containers aside from span to be used. This allows the inheriting 2d transformers to require matrix views be passed.
It is desirable to require matrix data to be passed with an associated view to ensure that any reshaping is explicit (though such functionality is not supported yet).