Static extent support
Added support for static extent matrices. Given matrices contain a span, static extent matrices are not trivially constructable because static extent spans are not trivially constructable.
Dynamic extent matrices are trivially constructable so there needs to be an empty/default constructor that is only available to dynamic extent matrices. This is accomplished using a static_assert.
Adding a new constructor to a specialization requires full template specializaiton which requires reimplementation of identical code or inheritance from a base template in this case. CRTP might be
able to avoid that, but CRTP seems overkill and static_assert within a template is what was used prior to concepts so it seems a reasonable solution. It also allows a custom error message to be given
which is helpful for troubleshooting complile errors.

Moving the YCbCr transformer to use static extent matrices for its transform matrices requires the matrices to be initialized in the initializer list. The functions to initialize the transform matrices
now get C_r, C_g, and C_b from the class fields instead of having them passed as explicit arguments. This avoids having to recompute r_g.