YCbCr inverse matrix
The inverse matrix is now correctly set. Wrote a test which multiplies the forward and inverse matrices to check the product against the identity matrix.
Added a basic multiply operation for the matrix type. It is currently not a cache optimized algorithm but for small matrices it does not matter much.