YCbCr transform test cases
Finished implementation the YCbCr transformer and it passes the currently written tests of the inverse matrix actually being the inverse of the transform matrix
and the channel transformer returning the original image (within tolerance) when applying the transform and then inverse in succession.

The test data was generated from a very simple python script to save the data from a png as a csv. csv was chosen because it is simple to write a parser for.

The previous const specification for the input of the YCbCr transformer did not work as desired because C++ does not coerce deep const in template substitution.
Matrix<T, ...> does not get coerced to Matrix<const T, ...> which makes calling functions which require const very annoying as a manual conversion of the matrix must take place before each function call.

Having the input data be const is desirable so finding a way to make this work is desirable. Perhaps an intermediate public method which coerces deep const and then calls a private method would work.
A seperate Abstract_Transformer for Matrix operations has been written for now with the end goal of eventually recombining it with the Abstract_Transformer written for 1D.