Owning matrix support
A matrix can now own its data. This was allowed by switching to using shared_ptr<vector<T>> instead of span<T>.
This change is motivated by the annoyance of returning new matrices for non-inplace operations e.g. multiply.
Having to declare space on the stack to store a matrix before multipling 2 matrices or having functions return a pair of the data and the matrix are both annoying to code with.

Constructors and methods for generating views are not yet added, but their addition should now be straightforward (just copy the shared_ptr).

Fixed an extraneous slash in the gitignore.

The double indirection of share_ptr<vector<T>> is inelegant e.g. (*this->data)[...] and shared_ptr<T[]> seems to have better syntax and perhaps performance.
The current use of vectors is because it provides support for a std::initializer_list and being able to construct a Matrix with an initializer list is highly desirable.
Currently, only 1D initalizer lists ares supported even though it is less error prone to write {{1, 2}, {3, 4}} than {1, 2, 3, 4}. This is another feature to consider.