Added readme.md
Added a readme.md to outline the project and implementation choices. Currently it is focused on multichannel matrices because it is valuable
to outline the code as it is being written to help keep ideas in order and provide an opportunity to review decisions. There is still significant
additions to be made and perhaps the file will be split in the future.