Multichannel matrix code moved to own file
Seperating declarations and implementations for Multichannel_Matrix added signficant bloat. The _impl was removed from the file to indicate
that it worked as a standalone header and was no longer just an implementation. The _impl was additionally removed from matrix_operations.hpp
as that file did not have its declarations seperately declared elsewhere.

Added more information to the readme.md

Improved comments and formatting for the Multichannel_Matrix code.

Renamed Image_View to Image_Matrix to indicate that it could be either owning or non-owning instead of exclusively non-owning.