Matrix transforms now coerce const on data
To enforce that certain operations do not change input data, added a public template function which accepts a Matrix_Like and then coerces into a Matrix with const data.
To support const data, a new template parameter was added for the container used to store the underlying data that way a const vector can be specified.

Abstract_Matrix_Transformer contains implementations of the public methods which forward the coerced const input to pure virtual implementation functions
which are to be implemented by inheriting classes.

To facillitate this, new concepts were added including Matrix_Like to enforce that the input can be coerced into a const data matrix. It is potentially worthwhile to
make the matrix shallow const in addition to deep const and have transform accept a const reference. The downside of passing as a const reference is that a reference
to a shared_ptr does not count as usage so pathological code could cause a transform to be applied to data that no longer exists (reference count hits zero). Given
the implementation method is private, perhaps it makes sense to have the public method own a copy of the shared_ptr for the duration of the transform and then pass
the matrix by reference only. Creating 1 copy is better than creating 2.

Concepts are currently organized haphazardly throughout the project based on what they might be loosely associated with. It would be good to find a more satisfying
way of organizing concepts.

The readme has been updated.

The .gitignore has been updated to be more flexible for ignoring build files generated when only building part of the project.