Multichannel_Matrix interface improvements
Removed all instances of the currently unused extent parameter in matrix templates and matrix transform templates.
Optimizations for static extent matrices is not planned anytime soon (if ever) so it is worth cleaning up the interface.

Greatly expanded the readme to include more detail on deep const coercion and the current design. Additionally added cpp formatting to code blocks.

Switched to using const references for private transform implementations so that the compiler will use the const version of overloaded functions.
If this was not done, all const overloads would have to have seperate names as in the previous commit. This also reduces the copies of shared_ptr
generated during a transform to 1 instead of 2.