Abstract matrix definition
Added an abstract type which multichannel matrices inherit from. The goal is to then be able to create different matrix types with greater differences
in their underlying storage by having pure virtual index methods for data access in the parent abstract matrix class. This currently works for the const
version, but the use of const vectors is causing this to fail for non-const indexing.

Working theory is that pure virtual functions are required to exist while normal class functions only need to exist if they get called. If true, this
would explain why the compiler allows the non-const index functions in multichannel matrices built on const vectors but not when pure virtual.

Additionally, cleaned up imports