Abstract matrix index method fix
Found a solution for abstract matrix index methods not supporting const vectors for underlying data.
If the passed container is const (i.e. const vector but not span<const T>) then Abstract_Contiguous_MCM<const T, Channels>
is inherited from instead of Abstract_Contiguous_MCM<T, Channels>. This was implemented using template metaprogramming to
create a template Propagate_Const<From, To> which will generate a type that is const if From is also const. Otherwise,
To remains unchanged.

Additionally, switched to using the array indexing operator[] when indexing with a single value because that is conceptually
different (it does not use strides) unlike index(i,j) and index(i,j,k)